add: orders crud

+++ db/migrate/20230704095238_create_orders.rb
@@ -3,7 +3,7 @@ class CreateOrders < ActiveRecord::Migration[7.0]
     create_table :orders do |t|
       t.date :date, default: -> { 'CURRENT_TIMESTAMP' }
       t.float :sub_total, default: 0
-      t.float :taxes, default: 10
+      t.integer :taxes, default: 10
       t.float :total, default: 0
       t.references :restaurant, null: false, foreign_key: true
